@@ -56,13 +56,14 @@
     "eslint-plugin-mocha": "^4.0.0",
     "eslint-plugin-react": "^5.2.2",
     "jest": "^20.0.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.2.1",
+    "prop-types": "^15.6.1",
+    "react": "^16.3.2",
     "release-script": "^1.0.2",
     "sinon": "^2.2.0",
     "sinon-chai": "^2.10.0"
   },
   "dependencies": {
+    "react-is": "^16.3.2",
     "warning": "^3.0.0"
   }
 }
